day02: Solve part 1

+++ src/puzzles/day01.rs
@@ -1,6 +1,6 @@
 use crate::RunError;
 
-pub fn main(part: u8, data: String) -> Result<i32, RunError> {
+pub fn main(part: u8, data: String) -> Result<usize, RunError> {
     let parsed_data = parse_data(data)?;
 
     match part {
@@ -10,18 +10,18 @@ pub fn main(part: u8, data: String) -> Result<i32, RunError> {
     }
 }
 
-fn parse_data(data: String) -> Result<Vec<i32>, RunError> {
+fn parse_data(data: String) -> Result<Vec<usize>, RunError> {
     let lines: Vec<&str> = data[..].split('\n').collect();
 
     match lines.iter()
-        .map(|x| x.trim().parse::<i32>())
+        .map(|x| x.trim().parse::<usize>())
         .collect() {
             Ok(parsed_data) => Ok(parsed_data),
             Err(e) => Err(RunError::ParseInt(e))
         }
 }
 
-fn part1(values: Vec<i32>) -> Result<i32, RunError> {
+fn part1(values: Vec<usize>) -> Result<usize, RunError> {
     // Find two entries that sum to 2020 and return their product.
     for i in values.iter() {
         for j in values.iter() {
@@ -35,7 +35,7 @@ fn part1(values: Vec<i32>) -> Result<i32, RunError> {
     Err(RunError::PartFailed)
 }
 
-fn part2(values: Vec<i32>) -> Result<i32, RunError> {
+fn part2(values: Vec<usize>) -> Result<usize, RunError> {
     // Find three entries that sum to 2020 and return their product.
     for i in values.iter() {
         for j in values.iter() {
@@ -64,7 +64,7 @@ mod tests {
     675
     1456";
 
-    static SAMPLE_DATA: &'static [i32] = &[1721,
+    static SAMPLE_DATA: &'static [usize] = &[1721,
     979,
     366,
     299,
